requirements.txt
alembic==1.4.3
attrs==20.2.0
bcrypt==3.2.0
Brotli==1.0.9
certifi==2020.6.20
cffi==1.14.4
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Compress==1.8.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lml==0.1.0
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
pyexcel==0.6.5
pyexcel-io==0.6.3
pyexcel-xls==0.6.1
pyexcel-xlsx==0.6.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
texttable==1.6.3
toml==0.10.2
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
